Move protocol files to db directory
