+++ fastagency/helpers.py
@@ -13,12 +13,12 @@ from fastagency.saas_app_generator import (
 )
 
 from .auth_token.auth import create_deployment_auth_token
-from .models.base import Model, ObjectReference
-from .models.registry import Registry
 
 # from fastagency.app import add_model
 # from .db.helpers import find_model_using_raw, get_db_connection, get_user
-from .protocols.prisma import PrismaProtocol
+from .db.prisma import PrismaProtocol
+from .models.base import Model, ObjectReference
+from .models.registry import Registry
 
 T = TypeVar("T", bound=Model)
