+++ fastagency/io/ionats.py
@@ -13,10 +13,11 @@ from faststream.nats import NatsMessage
 from nats.js import api
 from pydantic import BaseModel
 
+from ..db.prisma import PrismaProtocol
+
 # from ..db.helpers import find_model_using_raw
 from ..models.teams.multi_agent_team import MultiAgentTeam
 from ..models.teams.two_agent_teams import TwoAgentTeam
-from ..protocols.prisma import PrismaProtocol
 from .app import app, broker, stream  # noqa
 
 if TYPE_CHECKING:
