+++ fastagency/models/agents/base.py
@@ -7,7 +7,7 @@ from typing_extensions import TypeAlias
 from fastagency.openapi.client import Client
 
 # from ...db.helpers import find_model_using_raw
-from ...protocols.prisma import PrismaProtocol
+from ...db.prisma import PrismaProtocol
 from ..base import Model
 from ..registry import Registry
 from ..toolboxes.toolbox import ToolboxRef
