+++ fastagency/models/base.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel, Field, create_model, model_validator
 from typing_extensions import TypeAlias
 
 # from ..db.helpers import find_model_using_raw
-from ..protocols.prisma import PrismaProtocol
+from ..db.prisma import PrismaProtocol
 
 M = TypeVar("M", bound="Model")
